Add initial structure; comply with file naming rules
