+++ node/silkworm/downloader/messages/inbound_block_bodies.cpp
@@ -1,5 +1,5 @@
 /*
-   Copyright 2021 The Silkworm Authors
+   Copyright 2021-2022 The Silkworm Authors
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -14,23 +14,16 @@
    limitations under the License.
 */
 
-#include "InboundNewBlock.hpp"
+#include "inbound_block_bodies.hpp"
 
-#include <algorithm>
-
-#include <silkworm/common/cast.hpp>
 #include <silkworm/common/log.hpp>
-#include <silkworm/downloader/internals/random_number.hpp>
-#include <silkworm/downloader/rpc/send_message_by_id.hpp>
 
 namespace silkworm {
 
-InboundNewBlock::InboundNewBlock(const sentry::InboundMessage& msg, WorkingChain& wc, SentryClient& s)
+InboundBlockBodies::InboundBlockBodies(const sentry::InboundMessage& msg, WorkingChain& wc, SentryClient& s)
     : InboundMessage(), working_chain_(wc), sentry_(s) {
-    if (msg.id() != sentry::MessageId::NEW_BLOCK_66)
-        throw std::logic_error("InboundNewBlock received wrong InboundMessage");
-
-    reqId_ = RANDOM_NUMBER.generate_one();  // for trace purposes
+    if (msg.id() != sentry::MessageId::BLOCK_BODIES_66)
+        throw std::logic_error("InboundBlockBodies received wrong InboundMessage");
 
     peerId_ = string_from_H512(msg.peer_id());
 
@@ -40,16 +33,17 @@ InboundNewBlock::InboundNewBlock(const sentry::InboundMessage& msg, WorkingChain
     SILK_TRACE << "Received message " << *this;
 }
 
-void InboundNewBlock::execute() {
-    // todo: Erigon header-downloader apparently processes this message even if it is not in a fetching phase - do we
-    // need the same?
+void InboundBlockBodies::execute() {
+    using namespace std;
 
-    // todo: implement in the block-downloader
+    SILK_TRACE << "Processing message " << *this;
+
+    // todo: implement
 }
 
-uint64_t InboundNewBlock::reqId() const { return reqId_; }
+uint64_t InboundBlockBodies::reqId() const { return packet_.requestId; }
 
-std::string InboundNewBlock::content() const {
+std::string InboundBlockBodies::content() const {
     std::stringstream content;
     content << packet_;
     return content.str();
