+++ node/silkworm/downloader/messages/inbound_block_bodies.hpp
@@ -1,5 +1,5 @@
 /*
-   Copyright 2021 The Silkworm Authors
+   Copyright 2021-2022 The Silkworm Authors
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -14,33 +14,32 @@
    limitations under the License.
 */
 
-#ifndef SILKWORM_INBOUNDNEWBLOCKHASHES_HPP
-#define SILKWORM_INBOUNDNEWBLOCKHASHES_HPP
+#ifndef SILKWORM_INBOUND_BLOCK_BODIES_HPP
+#define SILKWORM_INBOUND_BLOCK_BODIES_HPP
 
-#include <silkworm/downloader/packets/NewBlockHashesPacket.hpp>
+#include <silkworm/downloader/packets/block_bodies_packet.hpp>
 
-#include "InboundMessage.hpp"
+#include "inbound_message.hpp"
 
 namespace silkworm {
 
-class InboundNewBlockHashes : public InboundMessage {
+class InboundBlockBodies : public InboundMessage {
   public:
-    InboundNewBlockHashes(const sentry::InboundMessage&, WorkingChain&, SentryClient&);
+    InboundBlockBodies(const sentry::InboundMessage& msg, WorkingChain&, SentryClient&);
 
-    std::string name() const override { return "InboundNewBlockHashes"; }
+    std::string name() const override { return "InboundBlockBodies"; }
     std::string content() const override;
     uint64_t reqId() const override;
 
     void execute() override;
 
   private:
-    std::string peerId_;
-    NewBlockHashesPacket packet_;
-    uint64_t reqId_;
+    PeerId peerId_;
+    BlockBodiesPacket66 packet_;
     WorkingChain& working_chain_;
     SentryClient& sentry_;
 };
 
 }  // namespace silkworm
 
-#endif  // SILKWORM_INBOUNDNEWBLOCKHASHES_HPP
+#endif  // SILKWORM_INBOUND_BLOCK_BODIES_HPP
