+++ node/silkworm/downloader/messages/inbound_get_block_headers.cpp
@@ -1,5 +1,5 @@
 /*
-   Copyright 2021 The Silkworm Authors
+   Copyright 2021-2022 The Silkworm Authors
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -14,19 +14,21 @@
    limitations under the License.
 */
 
-#include "InboundGetBlockBodies.hpp"
+#include "inbound_get_block_headers.hpp"
 
+#include <silkworm/common/cast.hpp>
 #include <silkworm/common/log.hpp>
-#include <silkworm/downloader/internals/body_retrieval.hpp>
-#include <silkworm/downloader/packets/BlockBodiesPacket.hpp>
+#include <silkworm/downloader/internals/header_retrieval.hpp>
+#include <silkworm/downloader/packets/block_headers_packet.hpp>
 #include <silkworm/downloader/rpc/send_message_by_id.hpp>
 
 namespace silkworm {
 
-InboundGetBlockBodies::InboundGetBlockBodies(const sentry::InboundMessage& msg, Db::ReadOnlyAccess db, SentryClient& s)
-    : InboundMessage(), db_(db), sentry_(s) {
-    if (msg.id() != sentry::MessageId::GET_BLOCK_BODIES_66) {
-        throw std::logic_error("InboundGetBlockBodies received wrong InboundMessage");
+InboundGetBlockHeaders::InboundGetBlockHeaders(const sentry::InboundMessage& msg, Db::ReadOnlyAccess db,
+                                               SentryClient& sentry)
+    : InboundMessage(), db_(db), sentry_(sentry) {
+    if (msg.id() != sentry::MessageId::GET_BLOCK_HEADERS_66) {
+        throw std::logic_error("InboundGetBlockHeaders received wrong InboundMessage");
     }
 
     peerId_ = string_from_H512(msg.peer_id());
@@ -37,29 +39,26 @@ InboundGetBlockBodies::InboundGetBlockBodies(const sentry::InboundMessage& msg, 
     SILK_TRACE << "Received message " << *this;
 }
 
-/*
- // ReplyBlockBodiesRLP is the eth/66 version of SendBlockBodiesRLP.
-func (p *Peer) ReplyBlockBodiesRLP(id uint64, bodies []rlp.RawValue) error {
-        // Not packed into BlockBodiesPacket to avoid RLP decoding
-        return p2p.Send(p.rw, BlockBodiesMsg, BlockBodiesRLPPacket66{
-                RequestId:            id,
-                BlockBodiesRLPPacket: bodies,
-        })
-}
- */
-void InboundGetBlockBodies::execute() {
+void InboundGetBlockHeaders::execute() {
     using namespace std;
 
     SILK_TRACE << "Processing message " << *this;
 
-    BodyRetrieval body_retrieval(db_);
+    HeaderRetrieval header_retrieval(db_);
 
-    BlockBodiesPacket66 reply;
+    BlockHeadersPacket66 reply;
     reply.requestId = packet_.requestId;
-    reply.request = body_retrieval.recover(packet_.request);
+    if (holds_alternative<Hash>(packet_.request.origin)) {
+        reply.request = header_retrieval.recover_by_hash(get<Hash>(packet_.request.origin), packet_.request.amount,
+                                                         packet_.request.skip, packet_.request.reverse);
+    } else {
+        reply.request =
+            header_retrieval.recover_by_number(get<BlockNum>(packet_.request.origin), packet_.request.amount,
+                                               packet_.request.skip, packet_.request.reverse);
+    }
 
     if (reply.request.empty()) {
-        log::Trace() << "[WARNING] Not replying to " << identify(*this) << ", no blocks found";
+        log::Trace() << "[WARNING] Not replying to " << identify(*this) << ", no headers found";
         return;
     }
 
@@ -67,13 +66,13 @@ void InboundGetBlockBodies::execute() {
     rlp::encode(rlp_encoding, reply);
 
     auto msg_reply = std::make_unique<sentry::OutboundMessageData>();
-    msg_reply->set_id(sentry::MessageId::BLOCK_BODIES_66);
+    msg_reply->set_id(sentry::MessageId::BLOCK_HEADERS_66);
     msg_reply->set_data(rlp_encoding.data(), rlp_encoding.length());  // copy
 
     SILK_TRACE << "Replying to " << identify(*this) << " using send_message_by_id with "
-                 << reply.request.size() << " bodies";
+                        << reply.request.size() << " headers";
 
-    rpc::SendMessageById rpc(peerId_, std::move(msg_reply));
+    rpc::SendMessageById rpc{peerId_, std::move(msg_reply)};
     rpc.do_not_throw_on_failure();
     sentry_.exec_remotely(rpc);
 
@@ -86,12 +85,11 @@ void InboundGetBlockBodies::execute() {
         SILK_TRACE << "Failure of rpc " << identify(*this) << ": "
                      << rpc.status().error_message();
     }
-
 }
 
-uint64_t InboundGetBlockBodies::reqId() const { return packet_.requestId; }
+uint64_t InboundGetBlockHeaders::reqId() const { return packet_.requestId; }
 
-std::string InboundGetBlockBodies::content() const {
+std::string InboundGetBlockHeaders::content() const {
     std::stringstream content;
     content << packet_;
     return content.str();
