+++ node/silkworm/downloader/messages/outbound_new_block_hashes.cpp
@@ -1,5 +1,5 @@
 /*
-Copyright 2021 The Silkworm Authors
+Copyright 2021-2022 The Silkworm Authors
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-#include "OutboundNewBlockHashes.hpp"
+#include "outbound_new_block_hashes.hpp"
 
 #include <silkworm/common/log.hpp>
 #include <silkworm/downloader/rpc/send_message_to_all.hpp>
