+++ node/silkworm/downloader/packets/get_block_bodies_packet.hpp
@@ -14,8 +14,8 @@
    limitations under the License.
 */
 
-#ifndef SILKWORM_GETBLOCKBODIESPACKET_HPP
-#define SILKWORM_GETBLOCKBODIESPACKET_HPP
+#ifndef SILKWORM_GET_BLOCK_BODIES_PACKET_HPP
+#define SILKWORM_GET_BLOCK_BODIES_PACKET_HPP
 
 #include <silkworm/downloader/internals/types.hpp>
 
@@ -55,4 +55,4 @@ inline std::ostream& operator<<(std::ostream& os, const GetBlockBodiesPacket66& 
 
 }  // namespace silkworm
 
-#endif  // SILKWORM_GETBLOCKBODIESPACKET_HPP
+#endif  // SILKWORM_GET_BLOCK_BODIES_PACKET_HPP
