+++ node/silkworm/downloader/packets/new_block_packet.hpp
@@ -1,5 +1,5 @@
 /*
-   Copyright 2021 The Silkworm Authors
+   Copyright 2021-2022 The Silkworm Authors
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -14,22 +14,34 @@
    limitations under the License.
 */
 
-#ifndef SILKWORM_MESSAGE_HPP
-#define SILKWORM_MESSAGE_HPP
+#ifndef SILKWORM_NEW_BLOCK_PACKET_HPP
+#define SILKWORM_NEW_BLOCK_PACKET_HPP
 
 #include <silkworm/downloader/internals/types.hpp>
 
 namespace silkworm {
 
-class Message {
-  public:
-    virtual std::string name() const = 0;
-
-    virtual void execute() = 0;  // inbound message send a reply, outbound message send a request
-
-    virtual ~Message() = default;
+struct NewBlockPacket {
+    Block block;
+    BigInt td;  // total difficulty
 };
+
+namespace rlp {
+
+    void encode(Bytes& to, const NewBlockPacket& from) noexcept;
+
+    size_t length(const NewBlockPacket& from) noexcept;
+
+    template <>
+    DecodingResult decode(ByteView& from, NewBlockPacket& to) noexcept;
+
+}  // namespace rlp
+
+inline std::ostream& operator<<(std::ostream& os, const NewBlockPacket& packet) {
+    os << "block num " << packet.block.header.number;
+    return os;
+}
 
 }  // namespace silkworm
 
-#endif  // SILKWORM_MESSAGE_HPP
+#endif  // SILKWORM_NEW_BLOCK_PACKET_HPP
