+++ node/silkworm/downloader/packets/packet_coding_test.cpp
@@ -16,11 +16,11 @@
 
 #include <catch2/catch.hpp>
 
-#include "BlockHeadersPacket.hpp"
-#include "GetBlockBodiesPacket.hpp"
-#include "GetBlockHeadersPacket.hpp"
-#include "NewBlockHashesPacket.hpp"
-#include "NewBlockPacket.hpp"
+#include "block_headers_packet.hpp"
+#include "get_block_bodies_packet.hpp"
+#include "get_block_headers_packet.hpp"
+#include "new_block_hashes_packet.hpp"
+#include "new_block_packet.hpp"
 
 // generic implementations (must follow types)
 #include <silkworm/rlp/encode_vector.hpp>
